requirements.txt
asgiref==3.4.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.2
defusedxml==0.7.1
Django==3.2
django-cors-headers==3.10.1
django-filter==21.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
drf-extensions==0.7.1
idna==3.4
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
mysqlclient==2.1.1
oauthlib==3.2.2
Pillow==8.4.0
pycparser==2.21
pycryptodome==3.17
PyJWT==1.7.1
PyMySQL==1.0.2
python3-openid==3.2.0
pytz==2022.7.1
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
typing-extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.15
